+++ packages/importer/src/config.ts
@@ -15,10 +15,6 @@ export interface OneSchemaConfig {
    */
   acceptCodeHookSuggestions?: boolean
   /**
-   * A token for a session created through the API for initiliaizing OneSchema
-   */
-  sessionToken?: string
-  /**
    * Options for specifying and/or modifying content displayed on different panes of the
    * OneSchema Importer
    */
@@ -112,9 +108,43 @@ export interface OneSchemaLaunchParams {
 }
 
 /**
- * Parameters for the OneSchema importer, includes all settings
+ * Parameters that can be set when the OneSchema importer launches with a sessionToken
  */
-export interface OneSchemaParams extends Partial<OneSchemaLaunchParams> {
+export interface OneSchemaLaunchSessionParams {
+  /**
+   * A token for a session created through the API for initiliaizing OneSchema
+   */
+  sessionToken: string
+  /**
+   * Config options for how the OneSchema importer will behave
+   */
+  config?: OneSchemaConfig
+}
+
+/**
+ * Possible errors when launching OneSchema
+ */
+export enum OneSchemaLaunchError {
+  MissingTemplate,
+  MissingJwt,
+  MissingSessionToken,
+}
+
+export interface OneSchemaLaunchStatus {
+  /**
+   * Whether or not launch was successful
+   */
+  success: boolean
+  /**
+   * If success is false, this will be why it failed
+   */
+  error?: OneSchemaLaunchError
+}
+
+/**
+ * Parameters for the OneSchema importer set at initialization
+ */
+export interface OneSchemaInitParams {
   /**
    * The client id from your OneSchema developer dashboard
    */
@@ -159,6 +189,18 @@ export interface OneSchemaParams extends Partial<OneSchemaLaunchParams> {
    */
   baseUrl?: string
 }
+
+/**
+ * Combined options for params used when launching OneSchema
+ */
+export type OneSchemaLaunchParamOptions =
+  | OneSchemaLaunchParams
+  | OneSchemaLaunchSessionParams
+
+/**
+ * Parameters for the OneSchema importer, includes all settings
+ */
+export type OneSchemaParams = OneSchemaInitParams & Partial<OneSchemaLaunchParamOptions>
 
 /**
  * The default values for the OneSchema importer