+++ packages/importer/src/importer.ts
@@ -1,6 +1,13 @@
 import { EventEmitter } from "eventemitter3"
 import merge from "lodash.merge"
-import { DEFAULT_PARAMS, OneSchemaLaunchParams, OneSchemaParams } from "./config"
+import {
+  DEFAULT_PARAMS,
+  OneSchemaLaunchError,
+  OneSchemaLaunchParams,
+  OneSchemaLaunchSessionParams,
+  OneSchemaLaunchStatus,
+  OneSchemaParams,
+} from "./config"
 
 const MAX_LAUNCH_RETRY = 10
 
@@ -120,9 +127,7 @@ export class OneSchemaImporterClass extends EventEmitter {
    * Launch will show the OneSchema window and initialize the importer session
    * @param launchParams optionally pass in parameter overrides or values not passed into constructor
    */
-  launch(launchParams?: Partial<OneSchemaLaunchParams>) {
-    window.addEventListener("message", this.#eventListener)
-
+  launch(launchParams?: Partial<OneSchemaLaunchParams>): OneSchemaLaunchStatus {
     const mergedParams = merge({}, this.#params, launchParams)
     const message: any = { messageType: "init" }
     message.manualClose = true
@@ -131,18 +136,48 @@ export class OneSchemaImporterClass extends EventEmitter {
     message.userJwt = mergedParams.userJwt
     if (!message.userJwt) {
       console.error("OneSchema config error: missing userJwt")
-      return
+      return { success: false, error: OneSchemaLaunchError.MissingJwt }
     }
 
     message.templateKey = mergedParams.templateKey
     if (!message.templateKey) {
       console.error("OneSchema config error: missing templateKey")
-      return
+      return { success: false, error: OneSchemaLaunchError.MissingTemplate }
     }
 
     if (mergedParams.webhookKey) {
       message.webhookKey = mergedParams.webhookKey
     }
+
+    this._launch(message)
+    return { success: true }
+  }
+
+  /**
+   * Launch session will show the OneSchema window and initialize the importer session
+   * with the given session token
+   * @param launchParams optionally pass in parameter overrides or values not passed into constructor
+   */
+  launchSession(
+    launchParams?: Partial<OneSchemaLaunchSessionParams>,
+  ): OneSchemaLaunchStatus {
+    const mergedParams = merge({}, this.#params, launchParams)
+    const message: any = { messageType: "init-session" }
+    message.manualClose = true
+    message.options = mergedParams.config
+
+    message.sessionToken = mergedParams.sessionToken
+    if (!message.sessionToken) {
+      console.error("OneSchema config error: missing sessionToken")
+      return { success: false, error: OneSchemaLaunchError.MissingSessionToken }
+    }
+
+    this._launch(message)
+    return { success: true }
+  }
+
+  _launch(message: any) {
+    window.addEventListener("message", this.#eventListener)
 
     const postInit = () => {
       this._hasCancelled = false